Added system to allow changing command files and reloading them
Using the command "reload_commands", the bot will re-read all the command files and import their new behaviours; allowing you to tweak the code of a command and reload it in the bot without having to restart it.

As a side-effect, all command files have been renamed to match their command name.